Features

+++ backend/donations/views.py
@@ -11,11 +11,22 @@ from rest_framework.decorators import api_view
 from rest_framework import status
 from django.contrib.gis.geos import Point
 from rest_framework.permissions import IsAuthenticated
+# from .permissions import AllowOwner
+
+from firebase_admin import credentials
+from firebase_admin import auth
+from django.conf import settings
+import firebase_admin
+
+
+# firebase_creds = credentials.Certificate(settings.FIREBASE_CONFIG)
+# firebase_app = firebase_admin.initialize_app(firebase_creds)
 
 
 
 class UserView(APIView):
     # permission_classes = [IsAuthenticated]
+    # permission_classes = [AllowOwner]
     def get_object(self, id):
         try:
             return UserProfile.objects.get(pk = id)
@@ -27,12 +38,15 @@ class UserView(APIView):
             user = self.get_object(id)
             serializer = UserSerializer(user)
         else:
+            
             users = UserProfile.objects.all()
             serializer = UserSerializer(users, many=True)
+        
         return Response(serializer.data)
 
     def post(self, request):
         serializer = UserSerializer(data=request.data)
+        print(request.user)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -40,15 +54,21 @@ class UserView(APIView):
 
     def put(self, request, id=None):
         user = self.get_object(id)
-        serializer = UserSerializer(user, data=request.data, partial=True )
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user and user.uid == request.user.uid:
+            serializer = UserSerializer(user, data=request.data, partial=True )
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED, )
     def delete(self, request, id=None):
         user = self.get_object(id)
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user and user.uid == request.user.uid:
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class ItemView(APIView):
@@ -62,7 +82,7 @@ class ItemView(APIView):
     def get(self, request, id=None):
         if id:
             item = self.get_item(id)
-            serializer = UserSerializer(item)
+            serializer = ItemSerializer(item)
         else:
             items = Item.objects.all()
             serializer = ItemSerializer(items, many=True)
@@ -78,32 +98,18 @@ class ItemView(APIView):
 
     def put(self, request, id=None):
         item = self.get_item(id)
-        serializer = ItemSerializer(item, data=request.data, partial=True )
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user and item.user == request.user.uid:
+            serializer = ItemSerializer(item, data=request.data, partial=True )
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
     def delete(self, request, id=None):
         item = self.get_item(id)
-        item.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-# from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter
-# from allauth.socialaccount.providers.oauth2.client import OAuth2Client
-# from dj_rest_auth.registration.views import SocialLoginView
-
-# class GoogleLogin(SocialLoginView): # if you want to use Authorization Code Grant, use this
-#     adapter_class = GoogleOAuth2Adapter
-#     callback_url = '1029888009386-fee8kup4cp300op9r2629qkga9pa81ia.apps.googleusercontent.com'
-#     client_class = OAuth2Client
-#     def post(self, request):
-#         print(request.data)
-#         serializer = UserSerializer(data=request.data)
-#         if serializer.is_valid():
-#             serializer.save()
-#             return Response(serializer.data, status=status.HTTP_201_CREATED)
-#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-# class GoogleLogin(SocialLoginView): # if you want to use Implicit Grant, use this
-#     adapter_class = GoogleOAuth2Adapter
+        if request.user and item.user == request.user.uid:
+            item.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
